Consolidate FileSystem.Handle with Internal.FileSystem.Handle
